Move the test for regex_match to tests
