+++ tests/test_dp.py
@@ -16,7 +16,8 @@ from algorithms.dp import (
     longest_increasing_subsequence_optimized2,
     int_divide,
     num_decodings, num_decodings2,
-    word_break
+    word_break,
+    Solution
 )
 
 
@@ -182,6 +183,53 @@ class TestWordBreak(unittest.TestCase):
         string = "redrum"
         dictionary = ["redd", "rum"]
         self.assertFalse(word_break(string, dictionary))
+
+class TestRegexSolution(unittest.TestCase):
+    def test_none_0(self):
+        s = ""
+        p = ""
+        self.assertTrue(Solution().is_match(s, p))
+
+    def test_none_1(self):
+        s = ""
+        p = "a"
+        self.assertFalse(Solution().is_match(s, p))
+
+    def test_no_symbol_equal(self):
+        s = "abcd"
+        p = "abcd"
+        self.assertTrue(Solution().is_match(s, p))
+
+    def test_no_symbol_not_equal_0(self):
+        s = "abcd"
+        p = "efgh"
+        self.assertFalse(Solution().is_match(s, p))
+
+    def test_no_symbol_not_equal_1(self):
+        s = "ab"
+        p = "abb"
+        self.assertFalse(Solution().is_match(s, p))
+
+    def test_symbol_0(self):
+        s = ""
+        p = "a*"
+        self.assertTrue(Solution().is_match(s, p))
+
+    def test_symbol_1(self):
+        s = "a"
+        p = "ab*"
+        self.assertTrue(Solution().is_match(s, p))
+
+    def test_symbol_2(self):
+        # E.g.
+        #   s a b b
+        # p 1 0 0 0
+        # a 0 1 0 0
+        # b 0 0 1 0
+        # * 0 1 1 1
+        s = "abb"
+        p = "ab*"
+        self.assertTrue(Solution().is_match(s, p))
 
 if __name__ == '__main__':
     unittest.main()
